:recycle: Move PlaceTableContainers and PlaceForm components to Home page folder
