+++ src/pages/Home/components/PlaceTable/PlaceTable.scss
@@ -1,4 +1,4 @@
-@import "../../stylesheets/variables";
+@import "../../../../stylesheets/variables";
 
 .place-table-container {
 	overflow-x: auto;
